new prompt, fix batch running script, rename openai to gpt
