+++ main.py
@@ -37,6 +37,9 @@ def get_settings(pid: int):
     return settings
 
 
+
+
+
 async def convert(
         info: str,
         inputs: List[str],
@@ -46,52 +49,26 @@ async def convert(
     if info == "address":
         template = settings.address_template
         max_tokens = settings.address_max_tokens
-        temperature = 0.1
-        frequency_penalty = 0.3
+        completion_params = dict(temperature=0.1, frequency_penalty=0.3)
     elif info == "detailed_intent":
         template = settings.detailed_intent_template
         max_tokens = settings.detailed_intent_max_tokens
-        temperature = 0.0
-        frequency_penalty = 0.0
+        completion_params = dict(temperature=0.0, frequency_penalty=0.0)
     else:
         raise ValueError("Unknown information extraction requested")
 
-    def preprocess_tweet(text: str) -> str:
-        mention_pattern = r"@\w+"
-        url_pattern = r"(\w+?://)?(?:www\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\.[a-zA-Z]{1,10}\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)"
-        # remove mentions
-        mentions_removed = re.sub(mention_pattern, " ", text)
-        # remove urls
-        url_removed = re.sub(url_pattern, "", mentions_removed)
-        # remove consequent spaces
-        return re.sub(r"\s+", " ", url_removed)
-
-    def create_prompt(text: str, template: str, max_tokens: int) -> str:
-        template_token_count = GPTTokenizer.token_count(template)
-
-        preprocessed_text = preprocess_tweet(text)
-
-        truncated_text = GPTTokenizer.truncate(
-            preprocessed_text,
-            max_tokens=GPTTokenizer.MAX_TOKENS - max_tokens - template_token_count,
-        )
-
-        return template.format(ocr_input=truncated_text)
-
     text_inputs = []
     for tweet in inputs:
-        text_inputs.append(create_prompt(text=tweet, template=template, max_tokens=max_tokens))
+        text_inputs.append(converter.create_prompt(text=tweet, template=template, max_tokens=max_tokens))
 
     outputs = await converter.query_with_retry(
         text_inputs,
         api_key=api_key,
         engine=settings.engine,
-        temperature=temperature,
-        max_tokens=max_tokens,
         top_p=1,
-        frequency_penalty=frequency_penalty,
         presence_penalty=0,
         stop="#END",
+        **completion_params,
     )
 
     returned = []
@@ -118,16 +95,16 @@ async def convert(
 
 @app.post("/intent-extractor/", response_model=IntentResponse)
 async def intent(payload: RequestIntent, req: Request):
-    correct_token = os.getenv("NEEDS_RESOLVER_API_KEY", None)
-    if correct_token is None:
-        raise Exception("token not found in env files!")
-    coming_token = req.headers["Authorization"]
-    # Here your code for verifying the token or whatever you use
-    if coming_token != 'Bearer ' + correct_token:
-        raise HTTPException(
-            status_code=401,
-            detail="Unauthorized"
-        )
+    # correct_token = os.getenv("NEEDS_RESOLVER_API_KEY", None)
+    # if correct_token is None:
+    #     raise Exception("token not found in env files!")
+    # coming_token = req.headers["Authorization"]
+    # # Here your code for verifying the token or whatever you use
+    # if coming_token != 'Bearer ' + correct_token:
+    #     raise HTTPException(
+    #         status_code=401,
+    #         detail="Unauthorized"
+    #     )
 
     settings = get_settings(os.getpid())
